refactor: remove duplicated code of db-specific TestContainers and RunnableSpecs
