+++ postgres/src/test/scala/zio/sql/postgresql/PostgresRunnableSpec.scala
@@ -1,36 +1,17 @@
 package zio.sql.postgresql
 
-import zio._
-import zio.test._
-import java.util.Properties
-import zio.sql.{ ConnectionPoolConfig, JdbcRunnableSpec }
+import com.dimafeng.testcontainers.{ JdbcDatabaseContainer, PostgreSQLContainer, SingleContainer }
+import org.testcontainers.utility.DockerImageName
+import zio.sql.JdbcRunnableSpec
 
 trait PostgresRunnableSpec extends JdbcRunnableSpec with PostgresJdbcModule {
 
-  def autoCommit: Boolean = true
-
-  private def connProperties(user: String, password: String): Properties = {
-    val props = new Properties
-    props.setProperty("user", user)
-    props.setProperty("password", password)
-    props
-  }
-
-  val poolConfigLayer = ZLayer.scoped {
-    TestContainer
-      .postgres()
-      .map(a =>
-        ConnectionPoolConfig(
-          url = a.jdbcUrl,
-          properties = connProperties(a.username, a.password),
-          autoCommit = autoCommit
-        )
-      )
-  }
-
-  override def spec: Spec[TestEnvironment, Any] =
-    specLayered.provideCustomShared(jdbcLayer)
-
-  def specLayered: Spec[JdbcEnvironment, Object]
+  override protected def getContainer: SingleContainer[_] with JdbcDatabaseContainer =
+    new PostgreSQLContainer(
+      dockerImageNameOverride = Option("postgres:alpine").map(DockerImageName.parse)
+    ).configure { a =>
+      a.withInitScript("db_schema.sql")
+      ()
+    }
 
 }
