+++ sqlserver/src/test/scala/zio/sql/sqlserver/SqlServerRunnableSpec.scala
@@ -1,34 +1,19 @@
 package zio.sql.sqlserver
 
-import zio._
-import zio.test._
-import java.util.Properties
-import zio.sql.{ ConnectionPoolConfig, JdbcRunnableSpec }
+import com.dimafeng.testcontainers.{ JdbcDatabaseContainer, MSSQLServerContainer, SingleContainer }
+import org.testcontainers.utility.DockerImageName
+import zio.sql.JdbcRunnableSpec
 
 trait SqlServerRunnableSpec extends JdbcRunnableSpec with SqlServerJdbcModule {
 
-  def autoCommit: Boolean = true
+  override protected def getContainer: SingleContainer[_] with JdbcDatabaseContainer =
+    new MSSQLServerContainer(
+      dockerImageName = DockerImageName
+        .parse("mcr.microsoft.com/azure-sql-edge:latest")
+        .asCompatibleSubstituteFor("mcr.microsoft.com/mssql/server")
+    ).configure { a =>
+      a.withInitScript("db_schema.sql")
+      ()
+    }
 
-  private def connProperties(user: String, password: String): Properties = {
-    val props = new Properties
-    props.setProperty("user", user)
-    props.setProperty("password", password)
-    props
-  }
-
-  val poolConfigLayer = ZLayer.scoped {
-    TestContainer.sqlServer
-      .map(a =>
-        ConnectionPoolConfig(
-          url = a.jdbcUrl,
-          properties = connProperties(a.username, a.password),
-          autoCommit = autoCommit
-        )
-      )
-  }
-
-  override def spec: Spec[TestEnvironment, Any] =
-    specLayered.provideCustomLayerShared(jdbcLayer)
-
-  def specLayered: Spec[JdbcEnvironment, Object]
 }
